Refactored adapter apis Moved Adapter APIs to suitables store actions Removed client usages from adpater APIs
