+++ src/stores/authStore.ts
@@ -88,7 +88,7 @@ export const useAuthStore = defineStore('authStore', {
         this.omsRedirectionUrl = payload.omsRedirectionUrl;
 
         const permissionId = process.env.VUE_APP_PERMISSION_ID;
-        const current = await useUserProfile().getProfile(this.token.value, this.getBaseUrl);
+        const current = await useUserProfile().getProfile();
         Settings.defaultZone = current.timeZone;
 
         const serverPermissionsFromRules = getServerPermissionsFromRules();
@@ -127,10 +127,10 @@ export const useAuthStore = defineStore('authStore', {
         });
         if (!facilities.length) throw "Unable to login. User is not associated with any facility"
 
-        await useProductStore().getFacilityPreference("SELECTED_FACILITY", current?.userId)
+        await useProductStore().getFacilityPreference("SELECTED_FACILITY")
         const currentFacility: any = useProductStore().getCurrentFacility
         isAdminUser ? await useProductStore().getDxpEComStores() : await useProductStore().getDxpEComStoresByFacility(currentFacility?.facilityId)
-        await useProductStore().getEComStorePreference("SELECTED_BRAND", current?.userId)
+        await useProductStore().getEComStorePreference("SELECTED_BRAND")
 
         setPermissions(appPermissions);
         // Fetch and set product identifier settings based on current product store
