+++ src/stores/productStore.ts
@@ -2,18 +2,7 @@ import { defineStore } from 'pinia'
 import api from '@/services/RemoteAPI'
 import { hasError } from '@/stores/authStore'
 import logger from '@/logger'
-import { useAuthStore } from './authStore'
 import { useProductMaster } from '@/composables/useProductMaster'
-import {
-  fetchGoodIdentificationTypes,
-  getEComStores,
-  getEComStoresByFacility,
-  getProductIdentificationPref,
-  getUserPreference,
-  setProductIdentificationPref,
-  setUserPreference,
-  getUserFacilities
-} from '@/adapter'
 import { useUserProfile } from './userProfileStore'
 import { showToast } from '@/services/uiUtils';
 import { translate } from '@/i18n'
@@ -82,9 +71,8 @@ export const useProductStore = defineStore('productStore', {
     },
 
     async getDxpEComStoresByFacility(facilityId?: any) {
-      const authStore = useAuthStore()
       try {
-        const response = await getEComStoresByFacility(authStore.token.value, authStore.getBaseUrl, 100, facilityId)
+        const response = await this.getEComStoresByFacility(100, facilityId)
         this.productStores = response
       } catch (error) {
         console.error(error)
@@ -93,9 +81,8 @@ export const useProductStore = defineStore('productStore', {
     },
 
     async getDxpEComStores() {
-      const authStore = useAuthStore()
       try {
-        const response = await getEComStores(authStore.token.value, authStore.getBaseUrl, 100)
+        const response = await this.getEComStores(100)
         this.productStores = response
       } catch (error) {
         console.error(error)
@@ -103,13 +90,12 @@ export const useProductStore = defineStore('productStore', {
       return this.productStores
     },
 
-    async getEComStorePreference(userPrefTypeId: any, userId = "") {
-      const authStore = useAuthStore()
+    async getEComStorePreference(userPrefTypeId: any) {
       if (!this.productStores.length) return
 
       let preferredStore = this.productStores[0]
       try {
-        const preferredStoreId = await getUserPreference(authStore.token.value, authStore.getBaseUrl, userPrefTypeId, userId)
+        const preferredStoreId = await useUserProfile().getUserPreference(userPrefTypeId)
         if (preferredStoreId) {
           const store = this.productStores.find((store: any) => store.productStoreId === preferredStoreId)
           if (store) preferredStore = store
@@ -121,12 +107,10 @@ export const useProductStore = defineStore('productStore', {
     },
 
     async setEComStorePreference(payload: any) {
-      const userProfile = useUserProfile().getUserProfile
       try {
-        await setUserPreference({
+        await useUserProfile().updateUserPreference({
           userPrefTypeId: 'SELECTED_BRAND',
           userPrefValue: payload.productStoreId,
-          userId: userProfile.userId
         })
       } catch (error) {
         console.error('error', error)
@@ -151,7 +135,7 @@ export const useProductStore = defineStore('productStore', {
 
       productIdentificationPref[id] = value
       try {
-        this.settings.productIdentifier.productIdentificationPref = await setProductIdentificationPref(
+        this.settings.productIdentifier.productIdentificationPref = await this.setProductIdentificationPref(
           eComStoreId,
           productIdentificationPref
         )
@@ -168,7 +152,7 @@ export const useProductStore = defineStore('productStore', {
         }
         return
       }
-      this.settings.productIdentifier.productIdentificationPref = await getProductIdentificationPref(eComStoreId)
+      this.settings.productIdentifier.productIdentificationPref = await this.getProductIdentificationPreference(eComStoreId)
     },
 
     async prepareProductIdentifierOptions() {
@@ -182,7 +166,7 @@ export const useProductStore = defineStore('productStore', {
         { goodIdentificationTypeId: 'title', description: 'Title' }
       ]
 
-      const fetchedGoodIdentificationTypes = await fetchGoodIdentificationTypes('HC_GOOD_ID_TYPE')
+      const fetchedGoodIdentificationTypes = await this.fetchGoodIdentificationTypes('HC_GOOD_ID_TYPE')
       const fetchedOptions = fetchedGoodIdentificationTypes || []
 
       this.settings.productIdentifier.productIdentificationOptions = Array.from(
@@ -320,9 +304,8 @@ export const useProductStore = defineStore('productStore', {
     },
 
     async getDxpUserFacilities(partyId: string, facilityGroupId: string, isAdminUser: boolean, payload = {}) {
-      const authStore = useAuthStore()
       try {
-        const response = await getUserFacilities(authStore.token.value, authStore.getBaseUrl, partyId, facilityGroupId, isAdminUser, payload)
+        const response = await this.getUserFacilities(partyId, facilityGroupId, isAdminUser, payload)
         this.facilities = response
       } catch (error) {
         console.error('Failed to fetch user facilities:', error)
@@ -334,13 +317,12 @@ export const useProductStore = defineStore('productStore', {
       this.currentFacility = facility
     },
 
-    async getFacilityPreference(userPrefTypeId: string, userId = '') {
-      const authStore = useAuthStore()
+    async getFacilityPreference(userPrefTypeId: string) {
       if (!this.facilities.length) return
 
       let preferredFacility = this.facilities[0]
       try {
-        const preferredFacilityId = await getUserPreference(authStore.token.value, authStore.getBaseUrl, userPrefTypeId, userId)
+        const preferredFacilityId = await useUserProfile().getUserPreference(userPrefTypeId)
         if (preferredFacilityId) {
           const facility = this.facilities.find((facility: any) => facility.facilityId === preferredFacilityId)
           if (facility) preferredFacility = facility
@@ -352,18 +334,354 @@ export const useProductStore = defineStore('productStore', {
     },
 
     async setFacilityPreference(payload: any) {
-      const userProfile = useUserProfile().getUserProfile
       try {
-        await setUserPreference({
+        await useUserProfile().updateUserPreference({
           userPrefTypeId: 'SELECTED_FACILITY',
           userPrefValue: payload.facilityId,
-          userId: userProfile.userId
         })
         this.currentFacility = payload
       } catch (error) {
         console.error('Failed to set facility preference', error)
       }
     },
+
+    async fetchGoodIdentificationTypes(parentTypeId = "HC_GOOD_ID_TYPE"): Promise <any> {
+      try {
+        const resp: any = await api({
+          url: "oms/goodIdentificationTypes",
+          method: "get",
+          params: {
+            parentTypeId,
+            pageSize: 50
+          }
+        });
+    
+        return Promise.resolve(resp.data)
+      } catch(error) {
+        return Promise.reject({
+          code: 'error',
+          message: 'Failed to fetch good identification types',
+          serverResponse: error
+        })
+      }
+    },
+    async getUserFacilities(partyId: string, facilityGroupId: any, isAdminUser = false, payload = {}) {
+      return await this.fetchFacilities(partyId, facilityGroupId, isAdminUser, payload)
+    },
+
+    async fetchFacilities(partyId: string, facilityGroupId: string, isAdminUser: boolean, payload: any): Promise <any> {
+      let facilityIds: Array<string> = [];
+      let filters: any = {};
+      let resp = {} as any
+
+      // Fetch the facilities associated with party
+      if(partyId) {
+        try {
+          resp = await this.fetchFacilitiesByParty(partyId)
+
+          facilityIds = resp.map((facility: any) => facility.facilityId);
+          if (!facilityIds.length) {
+            return Promise.reject({
+              code: 'error',
+              message: 'Failed to fetch user facilities',
+              serverResponse: resp.data
+            })
+          }
+        } catch(error) {
+          return Promise.reject({
+            code: 'error',
+            message: 'Failed to fetch user facilities',
+            serverResponse: error
+          })
+        }
+      }
+
+      // Fetch the facilities associated with group
+      if(facilityGroupId) {
+        try {
+          resp = await this.fetchFacilitiesByGroup(facilityGroupId, filters)
+
+          facilityIds = resp.map((facility: any) => facility.facilityId);
+          if (!facilityIds.length) {
+            return Promise.reject({
+              code: 'error',
+              message: 'Failed to fetch user facilities',
+              serverResponse: resp.data
+            })
+          }
+        } catch(error) {
+          return Promise.reject({
+            code: 'error',
+            message: 'Failed to fetch user facilities',
+            serverResponse: error
+          })
+        }
+      }
+
+      if(facilityIds.length) {
+        filters = {
+          facilityId: facilityIds.join(","),
+          facilityId_op: "in",
+          pageSize: facilityIds.length
+        }
+      }
+
+      const params = {
+        url: "oms/facilities",
+        method: "GET",
+        params: {
+          pageSize: 500,
+          ...payload,
+          ...filters
+        }
+      }
+
+      let facilities: Array<any> = []
+
+      try {
+        resp = await api(params);
+        facilities = resp.data
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to fetch facilities",
+          serverResponse: error
+        })
+      }
+
+      return Promise.resolve(facilities)
+    },
+
+    async fetchFacilitiesByGroup(facilityGroupId: string, payload?: any): Promise <any> {
+      const params = {
+        url: "oms/groupFacilities",
+        method: "GET",
+        params: {
+          facilityGroupId,
+          pageSize: 500,
+          ...payload
+        }
+      }
+    
+      let resp = {} as any;
+    
+      try {
+        resp = await api(params);    
+        // Filtering facilities on which thruDate is set, as we are unable to pass thruDate check in the api payload
+        // Considering that the facilities will always have a thruDate of the past.
+        const facilities = resp.data.filter((facility: any) => !facility.thruDate)
+        return Promise.resolve(facilities)
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to fetch facilities for group",
+          serverResponse: error
+        })
+      }
+    },
+    
+    async fetchFacilitiesByParty(partyId: string, payload?: any): Promise <Array<any> | Response> {
+      const params = {
+        url: `inventory-cycle-count/user/${partyId}/facilities`,
+        method: "GET",
+        params: {
+          ...payload,
+          pageSize: 500
+        }
+      }
+    
+      let resp = {} as any;
+    
+      try {
+        resp = await api(params);    
+        // Filtering facilities on which thruDate is set, as we are unable to pass thruDate check in the api payload
+        // Considering that the facilities will always have a thruDate of the past.
+        const facilities = resp.data.filter((facility: any) => !facility.thruDate)
+        return Promise.resolve(facilities)
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to fetch user associated facilities",
+          serverResponse: error
+        })
+      }
+    },
+
+    async getEComStores(pageSize = 100): Promise <any> {
+      const params = {
+        url: "oms/productStores",
+        method: "GET",
+        params: {
+          pageSize
+        }
+      }
+
+      let resp = {} as any;
+      let stores: Array<any> = []
+
+      try {
+        resp = await api(params);
+
+        stores = resp.data
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to fetch product stores",
+          serverResponse: error
+        })
+      }
+
+      return Promise.resolve(stores)
+    },
+
+    async getEComStoresByFacility(pageSize = 100, facilityId?: any): Promise <any> {
+      const params = {
+        url: `oms/facilities/${facilityId}/productStores`,
+        method: "GET",
+        params: {
+          pageSize,
+          facilityId
+        }
+      }
+
+      let resp = {} as any;
+      let stores: Array<any> = []
+
+      try {
+        resp = await api(params);
+
+        // Filtering stores on which thruDate is set, as we are unable to pass thruDate check in the api payload
+        // Considering that the stores will always have a thruDate of the past.
+        stores = resp.data.filter((store: any) => !store.thruDate)
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to fetch facility associated product stores",
+          serverResponse: error
+        })
+      }
+
+      if(!stores.length) return Promise.resolve(stores)
+
+      // Fetching all stores for the store name
+      const productStoresMap = {} as any;
+      try {
+        const productStores = await this.getEComStores(200);
+        productStores.map((store: any) => productStoresMap[store.productStoreId] = store.storeName)
+      } catch(error) {
+        console.error(error);
+      }
+
+      stores.map((store: any) => store.storeName = productStoresMap[store.productStoreId])
+      return Promise.resolve(stores)
+    },
+
+    async getProductIdentificationPreference(productStoreId: any): Promise<any> {
+      const productIdentifications = {
+        primaryId: "productId",
+        secondaryId: ""
+      }
+
+      try {
+        const resp = await api({
+          url: `oms/productStores/${productStoreId}/settings`,
+          method: "GET",
+          params: {
+            productStoreId,
+            settingTypeEnumId: "PRDT_IDEN_PREF"
+          }
+        }) as any;
+
+        const settings = resp.data
+        if(settings[0]?.settingValue) {
+          const respValue = JSON.parse(settings[0].settingValue)
+          productIdentifications['primaryId'] = respValue['primaryId']
+          productIdentifications['secondaryId'] = respValue['secondaryId']
+        } else {
+          await this.createProductIdentificationPref(productStoreId)
+        }
+      } catch(error: any) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to get product identification pref",
+          serverResponse: error
+        })
+      }
+
+      return productIdentifications;
+    },
+
+    async createProductIdentificationPref(productStoreId: string): Promise<any> {
+      const prefValue = {
+        primaryId: "productId",
+        secondaryId: ""
+      }
+
+      try {
+        await api({
+          url: `oms/productStores/${productStoreId}/settings`,
+          method: "POST",
+          data: {
+            productStoreId,
+            settingTypeEnumId: "PRDT_IDEN_PREF",
+            settingValue: JSON.stringify(prefValue)
+          }
+        });
+      } catch(err) {
+        console.error(err)
+      }
+
+      // not checking for resp success and fail case as every time we need to update the state with the
+      // default value when creating a pref
+      return prefValue;
+    },
+
+    async setProductIdentificationPref(productStoreId: string, productIdentificationPref: any): Promise<any> {
+      let resp = {} as any, isSettingExists = false;
+
+      try {
+        resp = await api({
+          url: `oms/productStores/${productStoreId}/settings`,
+          method: "GET",
+          params: {
+            productStoreId,
+            settingTypeEnumId: "PRDT_IDEN_PREF"
+          }
+        });
+
+        if(resp.data[0]?.settingTypeEnumId) isSettingExists = true
+      } catch(err) {
+        console.error(err)
+      }
+
+      if(!isSettingExists) {
+        return Promise.reject({
+          code: "error",
+          message: "product store setting is missing",
+          serverResponse: resp.data
+        })
+      }
+
+      try {
+        resp = await api({
+          url: `oms/productStores/${productStoreId}/settings`,
+          method: "POST",
+          data: {
+            productStoreId,
+            settingTypeEnumId: "PRDT_IDEN_PREF",
+            settingValue: JSON.stringify(productIdentificationPref)
+          }
+        });
+
+        return Promise.resolve(productIdentificationPref)
+      } catch(error) {
+        return Promise.reject({
+          code: "error",
+          message: "Failed to set product identification pref",
+          serverResponse: error
+        })
+      }
+    }
   },
 
   persist: true
